backend/test_sonar/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
factory-boy==3.2.1
Faker==15.1.1
PyJWT==2.5.0
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
